# docs/openapi.yaml
openapi: 3.0.3
info:
  title: GoRestTeach API
  description: |
    A teaching REST API built with **Go + Gin + GORM + PostgreSQL**.

    Designed to help Flutter teams learn how to work with REST endpoints and Postman.

    ## Authentication

    The API uses **JWT Bearer** authentication:
    1. Register via `POST /auth/register`
    2. Obtain tokens via `POST /auth/login`
    3. Add the header: `Authorization: Bearer <access_token>`
    4. The access token expires in **15 minutes**. Refresh it via `POST /auth/refresh`

    ## Response Format

    **Success:**
    ```json
    {
      "success": true,
      "data": { ... },
      "meta": { "page": 1, "per_page": 10, "total": 42 }
    }
    ```

    **Error:**
    ```json
    {
      "success": false,
      "error": {
        "code": "VALIDATION_ERROR",
        "message": "Validation failed",
        "details": [{ "field": "email", "message": "Must be a valid email address" }]
      }
    }
    ```

    ## Error Codes

    | HTTP | Code | When |
    |------|------|------|
    | 400 | `VALIDATION_ERROR` | Invalid request body fields |
    | 400 | `BAD_REQUEST` | Malformed JSON or missing body |
    | 401 | `UNAUTHORIZED` | Missing or invalid token |
    | 401 | `TOKEN_EXPIRED` | JWT access token has expired |
    | 403 | `FORBIDDEN` | Action not allowed (e.g. editing another user's post) |
    | 404 | `NOT_FOUND` | Resource does not exist |
    | 409 | `CONFLICT` | Email already registered |
...
  version: 1.0.0
  contact:
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Paste the `access_token` received from `/auth/login`.
        Format: `Bearer <token>`
  schemas:
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
              example: VALIDATION_ERROR
              enum:
                - VALIDATION_ERROR
                - BAD_REQUEST
                - UNAUTHORIZED
                - TOKEN_EXPIRED
                - FORBIDDEN
                - NOT_FOUND
                - CONFLICT
                - FILE_TOO_LARGE
                - UNSUPPORTED_MEDIA_TYPE
                - INTERNAL_ERROR
            message:
              type: string
              example: Validation failed
            details:
              type: array
              nullable: true
              items:
                type: object
                properties:
                  field:
                    type: string
                  message:
                    type: string
    UserPublic:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        name:
          type: string
          example: John Doe
        email:
          type: string
          format: email
          example: john@example.com
        bio:
          type: string
          example: Flutter developer at AcidSoft
        avatar_id:
          type: string
          format: uuid
          nullable: true
          example: 660e8400-e29b-41d4-a716-446655440001
          description: Use `GET /images/{avatar_id}` to fetch the image bytes
        created_at:
          type: string
          format: date-time
...
    Post:
      type: object
      properties:
        id:
          type: string
          format: uuid
        user_id:
          type: string
          format: uuid
        title:
          type: string
          example: My first post
        body:
          type: string
          example: This is the content of my first post.
        image_id:
          type: string
          format: uuid
          nullable: true
          description: Use `GET /images/{image_id}` to fetch the image bytes
        author:
          $ref: '#/components/schemas/UserPublic'
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 10
        total:
          type: integer
          example: 42
  responses:
    Unauthorized:
      description: Missing or invalid JWT token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: UNAUTHORIZED
              message: Authorization header is required
    Forbidden:
      description: User does not have permission to perform this action
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: FORBIDDEN
              message: You do not have permission to perform this action
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Post not found
    ValidationError:
      description: Request validation failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Validation failed
              details:
                - field: email
                  message: Must be a valid email address
                - field: password
                  message: Must be at least 8 characters long
paths:
...
  /posts:
    post:
      tags: [posts]
      summary: Create post
      description: |
        Creates a new post. The author is determined by the JWT token —
        you do **not** need to send a `user_id` field.

        **Team task (intermediate):** Create a post, then retrieve the full
        list and find your post by ID.
      operationId: createPost
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [title, body]
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: My Flutter experience
                body:
                  type: string
                  minLength: 10
                  example: Flutter is amazing for cross-platform development because...
      responses:
        '201':
          description: Post created
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'

    get:
      tags: [posts]
      summary: List posts
      description: |
        Returns a paginated list of posts with author information.

        Supports full-text case-insensitive search (`ILIKE`) on `title` and `body`.

        **Team task (intermediate):** Try different combinations of `page`,
        `per_page`, and `search`. Pay attention to the `meta` object in
        the response — use it to build pagination UI.
      operationId: listPosts
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
            minimum: 1
          example: 1
        - name: per_page
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
          example: 10
        - name: search
          in: query
          schema:
            type: string
          example: flutter
          description: Keyword matched against title and body (case-insensitive)
      responses:
        '200':
          description: Paginated list of posts
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Post'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /posts/{id}:
    get:
      tags: [posts]
      summary: Get post by ID
      description: Returns a single post with the full author object nested inside.
      operationId: getPostById
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Post details
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags: [posts]
      summary: Update post
      description: |
        Updates a post's `title` and/or `body`.
        **Only the post owner** can update it — other users receive `403 FORBIDDEN`.

        **Team task (advanced):** Create two separate user accounts. Log in as
        user A and create a post. Then switch to user B's token and try to
        update that post — you should receive `403 FORBIDDEN`.
      operationId: updatePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                  minLength: 3
                  maxLength: 255
                  example: Updated title
                body:
                  type: string
                  minLength: 10
                  example: Updated body content...
      responses:
        '200':
          description: Updated post
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags: [posts]
      summary: Delete post
      description: |
        Permanently deletes a post.
        **Only the post owner** can delete it.
      operationId: deletePost
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Deleted successfully (no body)
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /posts/{id}/image:
    post:
      tags: [posts]
      summary: Attach image to post
      description: |
        Uploads an image file and links it to the specified post.
        Only the post owner can attach images.

        After upload, `image_id` appears in the post object.
        Use `GET /images/{image_id}` to fetch and display the image.

        **Team task (advanced):** Attach an image to your post, then try
        uploading a `.txt` file instead — you should receive `415 UNSUPPORTED_MEDIA_TYPE`.
        Also verify that a different user cannot attach an image to someone
        else's post (`403 FORBIDDEN`).

        **Allowed types:** JPEG, PNG, WebP, GIF
        **Max size:** 5 MB
      operationId: attachPostImage
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Post UUID
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [image]
              properties:
                image:
                  type: string
                  format: binary
                  description: Image file (JPEG, PNG, WebP, or GIF — max 5 MB)
      responses:
        '200':
          description: Post updated with the attached image
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/SuccessResponse'
                  - type: object
                    properties:
                      data:
                        $ref: '#/components/schemas/Post'
        '400':
          description: Field `image` is missing
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '413':
          description: File exceeds the 5 MB limit
        '415':
          description: Unsupported file type

  # ── IMAGES ────────────────────────────────────────────────────────────────
...